# apps/api/src/swagger/api-schema.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/projects": {
      "get": {
        "description": "Fetches and returns an array of all project entities stored in the system.",
        "operationId": "ProjectsController_getProjects",
        "parameters": [],
        "responses": {
          "200": {
            "description": "List of all projects retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/GetProjectResponse"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error"
          },
          "503": {
            "description": "Service is unavailable"
          }
        },
        "summary": "Retrieve a list of all projects",
        "tags": [
          "Projects"
        ]
      },
      "post": {
        "description": "Creates a new project entity with the provided data. Requires permission: projects:create",
        "operationId": "ProjectsController_createProject",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProjectBody"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Project created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateProjectResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden, insufficient permissions"
          },
          "500": {
            "description": "Internal server error"
          },
          "503": {
            "description": "Service is unavailable"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a new project",
        "tags": [
          "Projects"
        ]
      }
    },
    "/projects/{uuid}": {
      "get": {
        "description": "Fetches and returns a single project entity based on the provided ID.",
        "operationId": "ProjectsController_getProjectById",
        "parameters": [
          {
            "name": "uuid",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Project retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetProjectResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid project UUID"
          },
          "404": {
            "description": "Project with the specified ID not found"
          },
          "500": {
            "description": "Internal server error"
          },
          "503": {
            "description": "Service is unavailable"
          }
        },
        "summary": "Retrieve a project",
        "tags": [
          "Projects"
        ]
      },
      "patch": {
        "description": "Updates the project entity identified by the provided ID with new data.  Requires permission: projects:update",
        "operationId": "ProjectsController_updateProject",
        "parameters": [
          {
            "name": "uuid",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProjectBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Project updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpdateProjectResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid project UUID"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden, insufficient permissions"
          },
          "404": {
            "description": "Project with the specified ID not found"
          },
          "500": {
            "description": "Internal server error"
          },
          "503": {
            "description": "Service is unavailable"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update an existing project",
        "tags": [
          "Projects"
        ]
      },
      "delete": {
        "description": "Deletes the project entity identified by the provided ID. Requires permission: projects:delete",
        "operationId": "ProjectsController_deleteProject",
        "parameters": [
          {
            "name": "uuid",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Project deleted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteProjectResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid project UUID"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden, insufficient permissions"
          },
          "404": {
            "description": "Project with the specified ID not found"
          },
          "500": {
            "description": "Internal server error"
          },
          "503": {
            "description": "Service is unavailable"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Delete a project",
        "tags": [
          "Projects"
        ]
      }
    },
...
    "/admin/projects/{uuid}": {
      "get": {
        "operationId": "AdminProjectsController_findOne",
        "parameters": [
          {
            "name": "uuid",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden, insufficient permissions"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Admin Projects"
        ]
      },
      "patch": {
        "operationId": "AdminProjectsController_update",
        "parameters": [
          {
            "name": "uuid",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProjectBody"
              }
            }
          }
        },
        "responses": {
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden, insufficient permissions"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Admin Projects"
        ]
      },
      "delete": {
        "operationId": "AdminProjectsController_remove",
        "parameters": [
          {
            "name": "uuid",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden, insufficient permissions"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Admin Projects"
        ]
      }
    },
...
  },
  "info": {
    "title": "Portfolio website API",
    "description": "API documentation for the portfolio website",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "ProjectStatusResponse": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "example": "IN_PROGRESS"
          },
          "label": {
            "type": "string",
            "example": "In progress"
          }
        },
        "required": [
          "code",
          "label"
        ]
      },
      "ProjectDeveloperResponse": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Patryk Z."
          },
          "role": {
            "type": "string",
            "example": "Full-Stack Developer"
          },
          "profileUrl": {
            "type": "string",
            "example": "https://github.com/Pieselak"
          }
        },
        "required": [
          "name"
        ]
      },
      "GetProjectResponse": {
        "type": "object",
        "properties": {
          "uuid": {
            "type": "string",
            "description": "Unique identifier for the project",
            "example": "123e4567-e89b-12d3-a456-426614174000"
          },
          "status": {
            "$ref": "#/components/schemas/ProjectStatusResponse"
          },
          "title": {
            "type": "string",
            "example": "Portfolio Website with Glucose Sensor Integration"
          },
          "shortDescription": {
            "type": "string",
            "example": "Modern portfolio website with CGM integration."
          },
          "detailedDescription": {
            "type": "string",
            "example": "The project presents a portfolio, projects module, and glucose sensor integration."
          },
          "imageUrl": {
            "type": "string",
            "example": "https://example.com/project-preview.png"
          },
          "tags": {
            "example": [
              "TypeScript",
              "React",
              "NestJS"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "sourceCodeOpen": {
            "type": "boolean",
            "example": true
          },
          "sourceCodeUrl": {
            "type": "string",
            "example": "https://github.com/Pieselak/Vite_React_ExpressJS_AboutMe"
          },
          "developers": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProjectDeveloperResponse"
            }
          },
          "startDate": {
            "type": "string",
            "example": "2025-11-19"
          },
          "completeDate": {
            "type": "string",
            "example": "2026-01-31"
          }
        },
        "required": [
          "uuid",
          "status",
          "title",
          "shortDescription",
          "detailedDescription",
          "tags",
          "sourceCodeOpen",
          "developers"
        ]
      },
      "TranslatedTextDto": {
        "type": "object",
        "properties": {
          "pl": {
            "type": "string"
          },
          "en": {
            "type": "string"
          },
          "de": {
            "type": "string"
          }
        },
        "required": [
          "pl",
          "en",
          "de"
        ]
      },
      "ProjectDeveloperBody": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Patryk Z."
          },
          "role": {
            "type": "string",
            "example": "Full-Stack Developer"
          },
          "profileUrl": {
            "type": "string",
            "example": "https://github.com/Pieselak"
          }
        },
        "required": [
          "name"
        ]
      },
      "CreateProjectBody": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "PLANNED",
              "IN_PROGRESS",
              "COMPLETED",
              "ON_HOLD"
            ],
            "example": "PLANNED"
          },
          "title": {
            "$ref": "#/components/schemas/TranslatedTextDto"
          },
          "shortDescription": {
            "$ref": "#/components/schemas/TranslatedTextDto"
          },
          "detailedDescription": {
            "$ref": "#/components/schemas/TranslatedTextDto"
          },
          "imageUrl": {
            "type": "string",
            "example": "https://example.com/project-preview.png"
          },
          "tags": {
            "example": [
              "TypeScript",
              "React",
              "NestJS"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "sourceCodeOpen": {
            "type": "boolean",
            "example": true
          },
          "sourceCodeUrl": {
            "type": "string",
            "example": "https://github.com/Pieselak/Vite_React_ExpressJS_AboutMe"
          },
          "developers": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProjectDeveloperBody"
            }
          },
          "startDate": {
            "type": "string",
            "example": "2025-11-19"
          },
          "completeDate": {
            "type": "string",
            "example": "2026-01-31"
          },
          "isPublished": {
            "type": "boolean",
            "default": false
          }
        },
        "required": [
          "status",
          "title",
          "shortDescription",
          "detailedDescription",
          "sourceCodeOpen"
        ]
      },
      "CreateProjectResponse": {
        "type": "object",
        "properties": {
          "uuid": {
            "type": "string",
            "description": "Unique identifier for the project",
            "example": "123e4567-e89b-12d3-a456-426614174000"
          },
          "status": {
            "$ref": "#/components/schemas/ProjectStatusResponse"
          },
          "title": {
            "type": "string",
            "example": "Portfolio Website with Glucose Sensor Integration"
          },
          "shortDescription": {
            "type": "string",
            "example": "Modern portfolio website with CGM integration."
          },
          "detailedDescription": {
            "type": "string",
            "example": "The project presents a portfolio, projects module, and glucose sensor integration."
          },
          "imageUrl": {
            "type": "string",
            "example": "https://example.com/project-preview.png"
          },
          "tags": {
            "example": [
              "TypeScript",
              "React",
              "NestJS"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "sourceCodeOpen": {
            "type": "boolean",
            "example": true
          },
          "sourceCodeUrl": {
            "type": "string",
            "example": "https://github.com/Pieselak/Vite_React_ExpressJS_AboutMe"
          },
          "developers": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProjectDeveloperResponse"
            }
          },
          "startDate": {
            "type": "string",
            "example": "2025-11-19"
          },
          "completeDate": {
            "type": "string",
            "example": "2026-01-31"
          }
        },
        "required": [
          "uuid",
          "status",
          "title",
          "shortDescription",
          "detailedDescription",
          "tags",
          "sourceCodeOpen",
          "developers"
        ]
      },
      "UpdateProjectBody": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "PLANNED",
              "IN_PROGRESS",
              "COMPLETED",
              "ON_HOLD"
            ],
            "example": "PLANNED"
          },
          "title": {
            "$ref": "#/components/schemas/TranslatedTextDto"
          },
          "shortDescription": {
            "$ref": "#/components/schemas/TranslatedTextDto"
          },
          "detailedDescription": {
            "$ref": "#/components/schemas/TranslatedTextDto"
          },
          "imageUrl": {
            "type": "string",
            "example": "https://example.com/project-preview.png"
          },
          "tags": {
            "example": [
              "TypeScript",
              "React",
              "NestJS"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "sourceCodeOpen": {
            "type": "boolean",
            "example": true
          },
          "sourceCodeUrl": {
            "type": "string",
            "example": "https://github.com/Pieselak/Vite_React_ExpressJS_AboutMe"
          },
          "developers": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProjectDeveloperBody"
            }
          },
          "startDate": {
            "type": "string",
            "example": "2025-11-19"
          },
          "completeDate": {
            "type": "string",
            "example": "2026-01-31"
          },
          "isPublished": {
            "type": "boolean",
            "default": false
          }
        }
      },
      "UpdateProjectResponse": {
        "type": "object",
        "properties": {
          "uuid": {
            "type": "string",
            "description": "Unique identifier for the project",
            "example": "123e4567-e89b-12d3-a456-426614174000"
          },
          "status": {
            "$ref": "#/components/schemas/ProjectStatusResponse"
          },
          "title": {
            "type": "string",
            "example": "Portfolio Website with Glucose Sensor Integration"
          },
          "shortDescription": {
            "type": "string",
            "example": "Modern portfolio website with CGM integration."
          },
          "detailedDescription": {
            "type": "string",
            "example": "The project presents a portfolio, projects module, and glucose sensor integration."
          },
          "imageUrl": {
            "type": "string",
            "example": "https://example.com/project-preview.png"
          },
          "tags": {
            "example": [
              "TypeScript",
              "React",
              "NestJS"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "sourceCodeOpen": {
            "type": "boolean",
            "example": true
          },
          "sourceCodeUrl": {
            "type": "string",
            "example": "https://github.com/Pieselak/Vite_React_ExpressJS_AboutMe"
          },
          "developers": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProjectDeveloperResponse"
            }
          },
          "startDate": {
            "type": "string",
            "example": "2025-11-19"
          },
          "completeDate": {
            "type": "string",
            "example": "2026-01-31"
          }
        },
        "required": [
          "uuid",
          "status",
          "title",
          "shortDescription",
          "detailedDescription",
          "tags",
          "sourceCodeOpen",
          "developers"
        ]
      },
      "DeleteProjectResponse": {
        "type": "object",
        "properties": {
          "deleted": {
            "type": "boolean",
            "example": true
          }
        },
        "required": [
          "deleted"
        ]
      },
...
    }
  }
}